% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR2.R
\name{SR2}
\alias{SR2}
\title{Create an An object with the provided parameters}
\usage{
SR2(
  data,
  confounders,
  split_ratio = 0.1,
  global_tol = 1e-08,
  sub_tol = 1e-05,
  tuning_iter = 30,
  max_iter = 50000
)
}
\arguments{
\item{split_ratio}{A proportion of elements will be drawn from data as testset, and the left over is trainset.}

\item{global_tol}{The global convergence criteria. One meet, the optimization will be terminated.}

\item{sub_tol}{The convergence creteria for elastic net problems, which decays with decreasing the loss of objective.}

\item{tuning_iter}{The number of iterations for each combination of tuning parameters}

\item{max_iter}{The maxiumme number of iterations. When it is reached, iteraction will terminated even if the global convergence criteria does not meet.}

\item{object}{An SC2 oject}
}
\value{
An SC2 object
}
\description{
Create an An object with the provided parameters
}
